requirements.txt
fastapi
uvicorn
pydantic
google-cloud-storage
faiss-cpu
numpy
vertexai
langchain-google-vertexai